Cast buffer cache size to uint64_t
* If we assign the cache size more than 2GB, cache size (4-byte int)
overflows and the cache grows infinitely.

Reviewed-on: http://review.couchbase.org/70539
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>

+++ src/filemgr.cc
@@ -395,7 +395,8 @@ void FileMgr::init(FileMgrConfig *config)
             // version supported.
             if (global_config.getNcacheBlock() > 0) {
                 if (ver_btreev2_format(ver_get_latest_magic())) {
-                    BnodeCacheMgr::init(global_config.getNcacheBlock() *
+                    BnodeCacheMgr::init(static_cast<uint64_t>(
+                                            global_config.getNcacheBlock()) *
                                         global_config.getBlockSize(),
                                         global_config.getFlushLimit());
                 } else {
